more work on controls page

+++ gui/app/controllers/controlsController.js
@@ -31,7 +31,6 @@
                 }
             ];
 
-
             $scope.gridUpdated = function() {
                 console.log("grid updated!");
                 console.log($scope.tiles);
@@ -46,10 +45,48 @@
                 return "No project selected";
             };
 
+            $scope.getScenesForSelectedProject = function() {
+                let currentProject = mixplayService.getCurrentProject();
+                if (currentProject != null) {
+                    return currentProject.scenes;
+                }
+                return [];
+            };
+
             $scope.deleteCurrentProject = function() {
                 let currentProject = mixplayService.getCurrentProject();
                 if (currentProject != null) {
-                    mixplayService.deleteProject(currentProject.id);
+
+                    utilityService
+                        .showConfirmationModal({
+                            title: "Delete MixPlay Project",
+                            question: `Are you sure you want to delete the MixPlay Project "${currentProject.name}"?`,
+                            confirmLabel: "Delete",
+                            confirmBtnType: "btn-danger"
+                        })
+                        .then(confirmed => {
+                            if (confirmed) {
+                                mixplayService.deleteProject(currentProject.id);
+                            }
+                        });
+                }
+            };
+
+            $scope.deleteScene = function(scene) {
+                if (scene != null) {
+
+                    utilityService
+                        .showConfirmationModal({
+                            title: "Delete Scene",
+                            question: `Are you sure you want to delete the Scene "${scene.name}"?`,
+                            confirmLabel: "Delete",
+                            confirmBtnType: "btn-danger"
+                        })
+                        .then(confirmed => {
+                            if (confirmed) {
+                                mixplayService.deleteSceneFromCurrentProject(scene.id);
+                            }
+                        });
                 }
             };
 
@@ -63,7 +100,7 @@
                         saveText: "Create",
                         validationFn: (value) => {
                             return new Promise(resolve => {
-                                if (value == null || value.length < 1) {
+                                if (value == null || value.trim().length < 1) {
                                     resolve(false);
                                 } else {
                                     resolve(true);
@@ -75,6 +112,34 @@
                     },
                     (name) => {
                         mixplayService.createNewProject(name);
+                    });
+            };
+
+            $scope.showCreateSceneModal = function() {
+
+                utilityService.openGetInputModal(
+                    {
+                        model: "",
+                        label: "New Scene Name",
+                        saveText: "Add",
+                        validationFn: (value) => {
+                            return new Promise(resolve => {
+                                if (value == null || value.trim().length < 1) {
+                                    resolve(false);
+                                } else {
+                                    resolve(true);
+                                }
+                            });
+                        },
+                        validationText: "Scene name cannot be empty."
+
+                    },
+                    (name) => {
+
+                        let currentProject = mixplayService.getCurrentProject();
+                        if (currentProject != null) {
+                            mixplayService.addNewSceneToCurrentProject(name);
+                        }
                     });
             };
         });
